refactor: consolidate PyTorch test runners into run_pytorch_tests.py
Address review feedback on the run_test.py migration: rename
run_pytorch_tests_full.py to run_pytorch_tests.py and delete the old
direct-pytest runner, leaving a single test-runner entrypoint.

- Update both "Test PyTorch Wheels" workflows to call the renamed script.
- The full workflow's Windows step keeps its historical 6-module subset
  via --include (the old runner hard-coded those modules).
- Update README, skip_tests docs, and code comments accordingly.
